# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZENITH Genesis Campaign API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
  code?: string;
  retryAfterSeconds?: number;
}

export interface NetworkConfig {
  chainId: string;
  chainIdDec: number;
  chainName: string;
  rpcUrl: string;
  blockExplorerUrl: string;
  currencySymbol: string;
  currencyName: string;
  decimals: number;
  totalSupply: string;
}

export interface GlobalStats {
  totalUsers: number;
  totalTransactions: number;
  totalZpAwarded: number;
  totalMissionsCompleted: number;
  totalFaucetClaims: number;
  totalBoxesOpened: number;
}

export type Tier = (typeof Tier)[keyof typeof Tier];

export const Tier = {
  bronze: "bronze",
  silver: "silver",
  gold: "gold",
  elite: "elite",
} as const;

export interface User {
  walletAddress: string;
  zp: number;
  as: number;
  rs: number;
  tier: Tier;
  tierMultiplier: number;
  finalScore: number;
  streakCount: number;
  lastCheckin?: string | null;
  lastClaim?: string | null;
  rank?: number | null;
  referralCode: string;
  invitedBy?: string | null;
  createdAt: string;
  txCount: number;
  uniqueContracts: number;
}

export interface ConnectRequest {
  address: string;
  referralCode?: string | null;
}

export type MissionType = (typeof MissionType)[keyof typeof MissionType];

export const MissionType = {
  basic: "basic",
  advanced: "advanced",
  social: "social",
} as const;

export interface Mission {
  id: number;
  slug: string;
  name: string;
  description: string;
  reward: number;
  type: MissionType;
  actionLabel?: string | null;
  actionUrl?: string | null;
}

export type MissionWithStatus = Mission & {
  completed: boolean;
  completedAt?: string | null;
};

export interface CompleteMissionRequest {
  address: string;
  missionSlug: string;
}

export interface CompleteMissionResponse {
  success: boolean;
  zpAwarded: number;
  user: User;
}

export interface FaucetRequest {
  address: string;
}

export interface FaucetResponse {
  success: boolean;
  amount: number;
  txHash: string;
  nextClaimAt: string;
}

export interface FaucetStatus {
  canClaim: boolean;
  nextClaimAt?: string | null;
  cooldownSecondsRemaining: number;
  lastTxHash?: string | null;
  amount: number;
}

export interface CheckinRequest {
  address: string;
}

export interface CheckinResponse {
  success: boolean;
  dayOfStreak: number;
  zpAwarded: number;
  bonusAwarded: number;
  nextCheckinAt: string;
  user: User;
}

export interface CheckinDay {
  day: number;
  label: string;
  reward: number;
  completed: boolean;
  isToday: boolean;
  isBonus: boolean;
}

export interface CheckinStatus {
  streakCount: number;
  canCheckin: boolean;
  nextCheckinAt?: string | null;
  cooldownSecondsRemaining: number;
  days: CheckinDay[];
}

export type BoxRarity = (typeof BoxRarity)[keyof typeof BoxRarity];

export const BoxRarity = {
  basic: "basic",
  rare: "rare",
  epic: "epic",
} as const;

export interface BoxConfig {
  rarity: BoxRarity;
  name: string;
  cost: number;
  minReward: number;
  maxReward: number;
  jackpot: number;
  jackpotChance: number;
  description: string;
  dailyLimit: number;
}

export interface OpenBoxRequest {
  address: string;
  rarity: BoxRarity;
}

export interface OpenBoxResponse {
  success: boolean;
  rarity: BoxRarity;
  zpReward: number;
  bonus?: string | null;
  isJackpot: boolean;
  user: User;
}

export interface LeaderboardEntry {
  rank: number;
  walletAddress: string;
  zp: number;
  tier: Tier;
  finalScore: number;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  you?: LeaderboardEntry | null;
  totalUsers: number;
}

export interface ReferralEntry {
  invitee: string;
  joinedAt: string;
  zpEarned: number;
}

export interface ReferralInfo {
  referralCode: string;
  referralLink: string;
  invitees: ReferralEntry[];
  totalInvited: number;
  totalZpEarned: number;
}

export interface RedeemReferralRequest {
  address: string;
  referralCode: string;
}

export interface RedeemReferralResponse {
  success: boolean;
  inviter: string;
  zpAwarded: number;
}

export interface ActivityLog {
  id: number;
  action: string;
  detail?: string | null;
  zpDelta: number;
  timestamp: string;
}

export interface Block {
  number: number;
  hash: string;
  timestamp: string;
  txCount: number;
  gasUsed: string;
  miner: string;
}

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  success: "success",
  failed: "failed",
  pending: "pending",
} as const;

export interface Transaction {
  hash: string;
  blockNumber: number;
  from: string;
  to: string;
  value: string;
  gasPrice: string;
  timestamp: string;
  status: TransactionStatus;
}

export type SearchResultKind =
  (typeof SearchResultKind)[keyof typeof SearchResultKind];

export const SearchResultKind = {
  address: "address",
  tx: "tx",
  block: "block",
  none: "none",
} as const;

export interface SearchResult {
  kind: SearchResultKind;
  block?: Block | null;
  transaction?: Transaction | null;
  address?: string | null;
  recentTransactions: Transaction[];
}

export type ListMissionsParams = {
  address?: string;
};

export type GetFaucetStatusParams = {
  address: string;
};

export type GetCheckinStatusParams = {
  address: string;
};

export type GetLeaderboardParams = {
  address?: string;
};

export type GetActivityFeedParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type GetLatestBlocksParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};

export type GetLatestTransactionsParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};

export type ExplorerSearchParams = {
  q: string;
};