# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZENITH Genesis Campaign API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /explorer/blocks:
    get:
      operationId: getLatestBlocks
      tags: [explorer]
      summary: Latest blocks on Zenith Testnet
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Block"
  /explorer/transactions:
    get:
      operationId: getLatestTransactions
      tags: [explorer]
      summary: Latest transactions on Zenith Testnet
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /explorer/search:
    get:
      operationId: explorerSearch
      tags: [explorer]
      summary: Search by address, tx hash, or block number
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"
components:
  schemas:
...
    Block:
      type: object
      properties:
        number:
          type: integer
        hash:
          type: string
        timestamp:
          type: string
          format: date-time
        txCount:
          type: integer
        gasUsed:
          type: string
        miner:
          type: string
      required: [number, hash, timestamp, txCount, gasUsed, miner]
    Transaction:
      type: object
      properties:
        hash:
          type: string
        blockNumber:
          type: integer
        from:
          type: string
        to:
          type: string
        value:
          type: string
        gasPrice:
          type: string
        timestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [success, failed, pending]
      required: [hash, blockNumber, from, to, value, gasPrice, timestamp, status]
    SearchResult:
      type: object
      properties:
        kind:
          type: string
          enum: [address, tx, block, none]
        block:
          $ref: "#/components/schemas/Block"
          nullable: true
        transaction:
          $ref: "#/components/schemas/Transaction"
          nullable: true
        address:
          type: string
          nullable: true
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required: [kind, recentTransactions]